Initial check-in
Espa_batch_worker.py is a script that is meant to run in
an AWS batch environment.  It is started with an order ID as
the only argument.  It will use the order ID to read an S3
object containing a JSON definition of a scene processing job
and will process the job.