Add batch queue submission code
Add the capability of submitting to the batch queue to
espa_submit.py.

Override parse_command_line() and add a --batch option.  Create
submit_SQS_job() for the existing submission to an SQS queue and
submit_batch_job() for submitting a batch job.  Rearrange code in
main() to use the new functions.