Change batch parameter value
Instead of just passing an order ID to espa-batch-worker, pass
an S3 URL with the bucket and key of the JSON processing order.
This reduces the amount of knowledge that espa-submit and
espa-batch-worker have to have in common.

Add a --queue command line option so the queue name can be
specified in the command line instead of requiring an
environment variable.  Espa-submit will look for an environment
variable SQSQueue if --queue is not specified.

While I'm here, clean up the way environment variables are used.