Change command-line parameter
Have espa-batch-worker look for an S3 URL as a command line
parameter rather than an order ID.  The URL specifies the bucket
and key of the JSON order file, so espa-batch-worker does not
need to know ahead of time what bucket to read.

Add a parse_s3_url() function to break down the URL.  Clean up
the imports.  Add new lines to the end of some error messages.