Add a batch command option to espa-submit
Add a command-line option "--batch-command" to espa_submit.py.
This allows a user to specify what script should be run in the
container when a batch job runs.

Also, allow a user to specify a value of "random" as an order ID
to request a random ID consisting of 4 lower-case alphabetic
characters followed by the string "-00001".  This is consistent
with the order IDs generated by submit_multi.py.

Finally, fix a bug when loading the order template from
the TEMPLATE_FILENAME specified in cli.py.