Add new espa-release-aux-info.spec file
The purpose of this new file is to keep track of the exact RPM versions
defined for aux scripts in a given ESPA release.  The deployer can use this
spec file to enforce a desired state on the Hadoop master node, as well as to
auto-update a release specification file (/etc/espa-release-aux).